Ref: moving peer helpers to a dedicated file.
