+++ src/common-helpers.ts
@@ -3,18 +3,14 @@ import { isHttpError } from "http-errors";
 import { createHash } from "node:crypto";
 import { z } from "zod";
 import { CommonConfig, InputSource, InputSources } from "./config-type";
-import {
-  InputValidationError,
-  MissingPeerError,
-  OutputValidationError,
-} from "./errors";
+import { InputValidationError, OutputValidationError } from "./errors";
 import { ZodFile } from "./file-schema";
 import { IOSchema } from "./io-schema";
+import { AbstractLogger } from "./logger";
 import { getMeta } from "./metadata";
 import { AuxMethod, Method } from "./method";
 import { mimeMultipart } from "./mime";
 import { ZodUpload } from "./upload-schema";
-import { AbstractLogger } from "./logger";
 
 export type FlatObject = Record<string, unknown>;
 
@@ -323,45 +319,3 @@ export type ErrMessage = Exclude<
 // the copy of the private Zod errorUtil.errToObj
 export const errToObj = (message: ErrMessage | undefined) =>
   typeof message === "string" ? { message } : message || {};
-
-export const loadPeer = async <T>(
-  moduleName: string,
-  moduleExport: string = "default",
-): Promise<T> => {
-  try {
-    return (await import(moduleName))[moduleExport];
-  } catch {}
-  try {
-    return await Promise.resolve().then(
-      /**
-       * alternative way for environments that do not support dynamic imports even it's CJS compatible
-       * @example jest with ts-jest
-       * @link https://github.com/evanw/esbuild/issues/2651
-       */
-      () => require(moduleName)[moduleExport],
-    );
-  } catch {}
-  throw new MissingPeerError(moduleName);
-};
-
-export const loadAltPeer = async <T>(
-  alternatives: { moduleName: string; moduleExport?: string }[],
-  fallback?: { moduleName: string; provider: () => T },
-) => {
-  for (const { moduleName, moduleExport } of alternatives) {
-    try {
-      return await loadPeer<T>(moduleName, moduleExport);
-    } catch {}
-  }
-  try {
-    const result = fallback?.provider();
-    if (result) {
-      return result;
-    }
-  } catch {}
-  throw new MissingPeerError(
-    alternatives
-      .concat(fallback ? [fallback] : [])
-      .map(({ moduleName }) => moduleName),
-  );
-};
